Implemented EventDao (wont be used)
Realized later that I can just use FirebaseFireStore and database to handle all my local database needs and will not require using Room. However, the experience is invaluable and is a good reference for job references

Moved around Repos and APIs to have a cleaner and clear responsibility as well as consistent structure.